[break] useCache, useResource now include entire response, rather than just entity
