+++ docs/api/useCache.md
@@ -8,27 +8,27 @@ title: useCache()
 ```typescript
 function useCache(
   fetchShape: ReadShape,
-  params: object | null
-): SchemaOf<typeof fetchShape.schema> | null;
+  params: object | null,
+): Denormalized<typeof fetchShape.schema> | null;
 ```
 
 <!--With Generics-->
 
 ```typescript
 function useCache<Params extends Readonly<object>, S extends Schema>(
-  { schema, getFetchKey }: ReadShape<S, Params>,
-  params: Params | null
-): SchemaOf<S> | null;
+  fetchShape: Pick<ReadShape<S, Params>, 'schema' | 'getFetchKey'>,
+  params: Params | null,
+): Denormalized<S> | null;
 ```
 
 <!--END_DOCUSAURUS_CODE_TABS-->
 
-> ### Rest Hooks 3.0 - Deprecation
+> ### Rest Hooks 3.0
 >
-> This hook is being deprecated in favor of [useCacheNew()](./useCacheNew)
+> This is the future default behavior of `useCache()` in version 3.0.
 >
-> - 3.0 this will be renamed to `useCacheLegacy()`
-> - 3.1 will remove `useCacheLegacy()`
+> - 3.0 will keep the legacy version as `useCacheLegacy()`
+> - 3.1 will remove both `useCacheLegacy()` and `useCache()`, leaving this behavior in `useCache()`
 
 Excellent to use data in the normalized cache without fetching.
 
@@ -41,7 +41,7 @@ Excellent to use data in the normalized cache without fetching.
 
 ## Example
 
-Using a type guard to deal with null
+### Using a type guard to deal with null
 
 ```tsx
 function Post({ id }: { id: number }) {
@@ -49,6 +49,40 @@ function Post({ id }: { id: number }) {
   // post as PostResource | null
   if (!post) return null;
   // post as PostResource (typeguarded)
+  // ...render stuff here
+}
+```
+
+### Paginated data
+
+When entities are stored in nested structures, that structure will remain.
+
+```typescript
+export class PaginatedPostResource extends Resource {
+  readonly id: number | null = null;
+  readonly title: string = '';
+  readonly content: string = '';
+
+  static urlRoot = 'http://test.com/post/';
+
+  static listShape<T extends typeof Resource>(this: T) {
+    return {
+      ...super.listShape(),
+      schema: { results: [this.getEntitySchema()], nextPage: '', lastPage: '' },
+    };
+  }
+}
+```
+
+```tsx
+function ArticleList({ page }: { page: string }) {
+  const { results: posts, nextPage, lastPage } = useCache(
+    PaginatedPostResource.listShape(),
+    { page },
+  );
+  // posts as PaginatedPostResource[] | null
+  if (!posts) return null;
+  // posts as PaginatedPostResource[] (typeguarded)
   // ...render stuff here
 }
 ```
