+++ docs/api/useResultCache.md
@@ -29,9 +29,9 @@ function useResultCache<Params extends Readonly<object>, D extends object>(
 
 > ### Rest Hooks 3.0 - Deprecation
 >
-> This hook is being deprecated in favor of [useCacheNew()](./useCacheNew)
+> This hook is being deprecated in favor of [useCache()](./useCache)
 >
-> - 3.0 `useCacheNew()` will be renamed to `useCache()`
+> - 3.0 `useCache()` will be renamed to `useCache()`
 > - 3.1 will remove `useResultCache()`
 
 Excellent to use with [pagination](../guides/pagination.md) or any other extra (non-entity) data in results.
