+++ src/index.ts
@@ -20,11 +20,11 @@ import reducer, { initialState } from './state/reducer';
 import { useDenormalized } from './state/selectors';
 import {
   useCache,
-  useCacheNew,
+  useCacheLegacy,
   useFetcher,
   useRetrieve,
   useResource,
-  useResourceNew,
+  useResourceLegacy,
   useSubscription,
   useResultCache,
   useMeta,
@@ -117,13 +117,13 @@ export {
   CacheProvider,
   ExternalCacheProvider,
   useCache,
-  useCacheNew,
+  useCacheLegacy,
   useFetcher,
   useRetrieve,
   useInvalidator,
   useResetter,
   useResource,
-  useResourceNew,
+  useResourceLegacy,
   useSubscription,
   useResultCache,
   useMeta,
