+++ src/react-integration/hooks/useCache.ts
@@ -2,7 +2,7 @@ import { useContext } from 'react';
 
 import { StateContext } from '~/react-integration/context';
 import { ReadShape, Schema } from '~/resource';
-import { useSchemaSelect } from '~/state/selectors';
+import { useDenormalized } from '~/state/selectors';
 
 /** Access a resource if it is available. */
 export default function useCache<
@@ -13,5 +13,5 @@ export default function useCache<
   params: Params | null,
 ) {
   const state = useContext(StateContext);
-  return useSchemaSelect(fetchShape, params, state);
+  return useDenormalized(fetchShape, params, state)[0];
 }
