+++ src/react-integration/hooks/useResource.ts
@@ -1,8 +1,10 @@
-import { ReadShape, Schema, SchemaOf } from '~/resource';
-import useCache from './useCache';
+import { ReadShape, Schema } from '~/resource';
+import { Denormalized, DenormalizedNullable } from '~/resource/normal';
+import { useDenormalized } from '~/state/selectors';
 import useRetrieve from './useRetrieve';
 import useError from './useError';
-import { useMemo } from 'react';
+import { useMemo, useContext } from 'react';
+import { StateContext } from '~/react-integration/context';
 
 import hasUsableData from './hasUsableData';
 
@@ -15,31 +17,30 @@ type ResourceArgs<S extends Schema, Params extends Readonly<object>> = [
 function useOneResource<Params extends Readonly<object>, S extends Schema>(
   fetchShape: ReadShape<S, Params>,
   params: Params | null,
-): CondNull<typeof params, NonNullable<SchemaOf<S>>> {
-  // maybePromise is undefined when data is stale or params is null
+): CondNull<typeof params, DenormalizedNullable<S>, Denormalized<S>> {
   const maybePromise = useRetrieve(fetchShape, params);
-  // resource is null when it is not in cache or params is null
-  const resource = useCache(fetchShape, params);
-  const error = useError(fetchShape, params, !!resource);
+  const state = useContext(StateContext);
+  const [denormalized, ready] = useDenormalized(fetchShape, params, state);
+  const error = useError(fetchShape, params, ready);
 
-  if (!hasUsableData(!!resource, fetchShape) && maybePromise)
-    throw maybePromise;
+  if (!hasUsableData(ready, fetchShape) && maybePromise) throw maybePromise;
   if (error) throw error;
 
-  return resource as any;
+  return denormalized as any;
 }
 
 /** many form resource */
 function useManyResources<A extends ResourceArgs<any, any>[]>(
   ...resourceList: A
 ) {
-  const resources = resourceList.map(
+  const state = useContext(StateContext);
+  const denormalizedValues = resourceList.map(
     <Params extends Readonly<object>, S extends Schema>([
       fetchShape,
       params,
     ]: ResourceArgs<S, Params>) =>
       // eslint-disable-next-line react-hooks/rules-of-hooks
-      useCache(fetchShape, params),
+      useDenormalized(fetchShape, params, state),
   );
   const promises = resourceList
     .map(([fetchShape, params]) =>
@@ -47,14 +48,17 @@ function useManyResources<A extends ResourceArgs<any, any>[]>(
       useRetrieve(fetchShape, params),
     )
     // only wait on promises without results
-    .map((p, i) => !hasUsableData(!!resources[i], resourceList[i][0]) && p);
+    .map(
+      (p, i) =>
+        !hasUsableData(denormalizedValues[i][1], resourceList[i][0]) && p,
+    );
 
   // throw first valid error
   for (let i = 0; i < resourceList.length; i++) {
     const [fetchShape, params] = resourceList[i];
-    const resource = resources[i];
+    const [_, ready] = denormalizedValues[i];
     // eslint-disable-next-line react-hooks/rules-of-hooks
-    const error = useError(fetchShape, params, !!resource);
+    const error = useError(fetchShape, params, ready);
     if (error && !promises[i]) throw error;
   }
 
@@ -68,23 +72,26 @@ function useManyResources<A extends ResourceArgs<any, any>[]>(
 
   if (promise) throw promise;
 
-  return resources;
+  return denormalizedValues.map(([denormalized]) => denormalized);
 }
 
-type CondNull<P, R> = P extends null ? null : R;
+type CondNull<P, A, B> = P extends null ? A : B;
 
 /** Ensure a resource is available; suspending to React until it is. */
 export default function useResource<
   P extends Readonly<object> | null,
+  B extends Readonly<object | string> | void,
   S extends Schema
 >(
   fetchShape: ReadShape<S, NonNullable<P>>,
   params: P,
-): CondNull<P, SchemaOf<S>>;
+): CondNull<P, DenormalizedNullable<S>, Denormalized<S>>;
 export default function useResource<
   P1 extends Readonly<object> | null,
   S1 extends Schema
->(v1: [ReadShape<S1, NonNullable<P1>>, P1]): [CondNull<P1, SchemaOf<S1>>];
+>(
+  v1: [ReadShape<S1, NonNullable<P1>>, P1],
+): [CondNull<P1, DenormalizedNullable<S1>, Denormalized<S1>>];
 export default function useResource<
   P1 extends Readonly<object> | null,
   S1 extends Schema,
@@ -93,7 +100,10 @@ export default function useResource<
 >(
   v1: [ReadShape<S1, NonNullable<P1>>, P1],
   v2: [ReadShape<S2, NonNullable<P2>>, P2],
-): [CondNull<P1, SchemaOf<S1>>, CondNull<P2, SchemaOf<S2>>];
+): [
+  CondNull<P1, DenormalizedNullable<S1>, Denormalized<S1>>,
+  CondNull<P2, DenormalizedNullable<S2>, Denormalized<S2>>,
+];
 export default function useResource<
   P1 extends Readonly<object> | null,
   S1 extends Schema,
@@ -106,9 +116,9 @@ export default function useResource<
   v2: [ReadShape<S2, NonNullable<P2>>, P2],
   v3: [ReadShape<S3, NonNullable<P3>>, P3],
 ): [
-  CondNull<P1, SchemaOf<S1>>,
-  CondNull<P2, SchemaOf<S2>>,
-  CondNull<P3, SchemaOf<S3>>,
+  CondNull<P1, DenormalizedNullable<S1>, Denormalized<S1>>,
+  CondNull<P2, DenormalizedNullable<S2>, Denormalized<S2>>,
+  CondNull<P3, DenormalizedNullable<S3>, Denormalized<S3>>,
 ];
 export default function useResource<
   P1 extends Readonly<object> | null,
@@ -125,10 +135,10 @@ export default function useResource<
   v3: [ReadShape<S3, NonNullable<P3>>, P3],
   v4: [ReadShape<S4, NonNullable<P4>>, P4],
 ): [
-  CondNull<P1, SchemaOf<S1>>,
-  CondNull<P2, SchemaOf<S2>>,
-  CondNull<P3, SchemaOf<S3>>,
-  CondNull<P4, SchemaOf<S4>>,
+  CondNull<P1, DenormalizedNullable<S1>, Denormalized<S1>>,
+  CondNull<P2, DenormalizedNullable<S2>, Denormalized<S2>>,
+  CondNull<P3, DenormalizedNullable<S3>, Denormalized<S3>>,
+  CondNull<P4, DenormalizedNullable<S4>, Denormalized<S4>>,
 ];
 export default function useResource<
   P1 extends Readonly<object> | null,
@@ -148,11 +158,11 @@ export default function useResource<
   v4: [ReadShape<S4, NonNullable<P4>>, P4],
   v5: [ReadShape<S5, NonNullable<P5>>, P5],
 ): [
-  CondNull<P1, SchemaOf<S1>>,
-  CondNull<P2, SchemaOf<S2>>,
-  CondNull<P3, SchemaOf<S3>>,
-  CondNull<P4, SchemaOf<S4>>,
-  CondNull<P5, SchemaOf<S5>>,
+  CondNull<P1, DenormalizedNullable<S1>, Denormalized<S1>>,
+  CondNull<P2, DenormalizedNullable<S2>, Denormalized<S2>>,
+  CondNull<P3, DenormalizedNullable<S3>, Denormalized<S3>>,
+  CondNull<P4, DenormalizedNullable<S4>, Denormalized<S4>>,
+  CondNull<P5, DenormalizedNullable<S5>, Denormalized<S5>>,
 ];
 export default function useResource<
   Params extends Readonly<object>,
