+++ src/react-integration/hooks/useResourceLegacy.ts
@@ -1,10 +1,8 @@
-import { ReadShape, Schema } from '~/resource';
-import { Denormalized, DenormalizedNullable } from '~/resource/normal';
-import { useDenormalized } from '~/state/selectors';
+import { ReadShape, Schema, SchemaOf } from '~/resource';
+import useCacheLegacy from './useCacheLegacy';
 import useRetrieve from './useRetrieve';
 import useError from './useError';
-import { useMemo, useContext } from 'react';
-import { StateContext } from '~/react-integration/context';
+import { useMemo } from 'react';
 
 import hasUsableData from './hasUsableData';
 
@@ -17,30 +15,31 @@ type ResourceArgs<S extends Schema, Params extends Readonly<object>> = [
 function useOneResource<Params extends Readonly<object>, S extends Schema>(
   fetchShape: ReadShape<S, Params>,
   params: Params | null,
-): CondNull<typeof params, DenormalizedNullable<S>, Denormalized<S>> {
+): CondNull<typeof params, NonNullable<SchemaOf<S>>> {
+  // maybePromise is undefined when data is stale or params is null
   const maybePromise = useRetrieve(fetchShape, params);
-  const state = useContext(StateContext);
-  const [denormalized, ready] = useDenormalized(fetchShape, params, state);
-  const error = useError(fetchShape, params, ready);
+  // resource is null when it is not in cache or params is null
+  const resource = useCacheLegacy(fetchShape, params);
+  const error = useError(fetchShape, params, !!resource);
 
-  if (!hasUsableData(ready, fetchShape) && maybePromise) throw maybePromise;
+  if (!hasUsableData(!!resource, fetchShape) && maybePromise)
+    throw maybePromise;
   if (error) throw error;
 
-  return denormalized as any;
+  return resource as any;
 }
 
 /** many form resource */
 function useManyResources<A extends ResourceArgs<any, any>[]>(
   ...resourceList: A
 ) {
-  const state = useContext(StateContext);
-  const denormalizedValues = resourceList.map(
+  const resources = resourceList.map(
     <Params extends Readonly<object>, S extends Schema>([
       fetchShape,
       params,
     ]: ResourceArgs<S, Params>) =>
       // eslint-disable-next-line react-hooks/rules-of-hooks
-      useDenormalized(fetchShape, params, state),
+      useCacheLegacy(fetchShape, params),
   );
   const promises = resourceList
     .map(([fetchShape, params]) =>
@@ -48,17 +47,14 @@ function useManyResources<A extends ResourceArgs<any, any>[]>(
       useRetrieve(fetchShape, params),
     )
     // only wait on promises without results
-    .map(
-      (p, i) =>
-        !hasUsableData(denormalizedValues[i][1], resourceList[i][0]) && p,
-    );
+    .map((p, i) => !hasUsableData(!!resources[i], resourceList[i][0]) && p);
 
   // throw first valid error
   for (let i = 0; i < resourceList.length; i++) {
     const [fetchShape, params] = resourceList[i];
-    const [_, ready] = denormalizedValues[i];
+    const resource = resources[i];
     // eslint-disable-next-line react-hooks/rules-of-hooks
-    const error = useError(fetchShape, params, ready);
+    const error = useError(fetchShape, params, !!resource);
     if (error && !promises[i]) throw error;
   }
 
@@ -72,27 +68,24 @@ function useManyResources<A extends ResourceArgs<any, any>[]>(
 
   if (promise) throw promise;
 
-  return denormalizedValues.map(([denormalized]) => denormalized);
+  return resources;
 }
 
-type CondNull<P, A, B> = P extends null ? A : B;
+type CondNull<P, R> = P extends null ? null : R;
 
 /** Ensure a resource is available; suspending to React until it is. */
-export default function useResourceNew<
+export default function useResourceLegacy<
   P extends Readonly<object> | null,
-  B extends Readonly<object | string> | void,
   S extends Schema
 >(
   fetchShape: ReadShape<S, NonNullable<P>>,
   params: P,
-): CondNull<P, DenormalizedNullable<S>, Denormalized<S>>;
-export default function useResourceNew<
+): CondNull<P, SchemaOf<S>>;
+export default function useResourceLegacy<
   P1 extends Readonly<object> | null,
   S1 extends Schema
->(
-  v1: [ReadShape<S1, NonNullable<P1>>, P1],
-): [CondNull<P1, DenormalizedNullable<S1>, Denormalized<S1>>];
-export default function useResourceNew<
+>(v1: [ReadShape<S1, NonNullable<P1>>, P1]): [CondNull<P1, SchemaOf<S1>>];
+export default function useResourceLegacy<
   P1 extends Readonly<object> | null,
   S1 extends Schema,
   P2 extends Readonly<object> | null,
@@ -100,11 +93,8 @@ export default function useResourceNew<
 >(
   v1: [ReadShape<S1, NonNullable<P1>>, P1],
   v2: [ReadShape<S2, NonNullable<P2>>, P2],
-): [
-  CondNull<P1, DenormalizedNullable<S1>, Denormalized<S1>>,
-  CondNull<P2, DenormalizedNullable<S2>, Denormalized<S2>>,
-];
-export default function useResourceNew<
+): [CondNull<P1, SchemaOf<S1>>, CondNull<P2, SchemaOf<S2>>];
+export default function useResourceLegacy<
   P1 extends Readonly<object> | null,
   S1 extends Schema,
   P2 extends Readonly<object> | null,
@@ -116,11 +106,11 @@ export default function useResourceNew<
   v2: [ReadShape<S2, NonNullable<P2>>, P2],
   v3: [ReadShape<S3, NonNullable<P3>>, P3],
 ): [
-  CondNull<P1, DenormalizedNullable<S1>, Denormalized<S1>>,
-  CondNull<P2, DenormalizedNullable<S2>, Denormalized<S2>>,
-  CondNull<P3, DenormalizedNullable<S3>, Denormalized<S3>>,
+  CondNull<P1, SchemaOf<S1>>,
+  CondNull<P2, SchemaOf<S2>>,
+  CondNull<P3, SchemaOf<S3>>,
 ];
-export default function useResourceNew<
+export default function useResourceLegacy<
   P1 extends Readonly<object> | null,
   S1 extends Schema,
   P2 extends Readonly<object> | null,
@@ -135,12 +125,12 @@ export default function useResourceNew<
   v3: [ReadShape<S3, NonNullable<P3>>, P3],
   v4: [ReadShape<S4, NonNullable<P4>>, P4],
 ): [
-  CondNull<P1, DenormalizedNullable<S1>, Denormalized<S1>>,
-  CondNull<P2, DenormalizedNullable<S2>, Denormalized<S2>>,
-  CondNull<P3, DenormalizedNullable<S3>, Denormalized<S3>>,
-  CondNull<P4, DenormalizedNullable<S4>, Denormalized<S4>>,
+  CondNull<P1, SchemaOf<S1>>,
+  CondNull<P2, SchemaOf<S2>>,
+  CondNull<P3, SchemaOf<S3>>,
+  CondNull<P4, SchemaOf<S4>>,
 ];
-export default function useResourceNew<
+export default function useResourceLegacy<
   P1 extends Readonly<object> | null,
   S1 extends Schema,
   P2 extends Readonly<object> | null,
@@ -158,15 +148,14 @@ export default function useResourceNew<
   v4: [ReadShape<S4, NonNullable<P4>>, P4],
   v5: [ReadShape<S5, NonNullable<P5>>, P5],
 ): [
-  CondNull<P1, DenormalizedNullable<S1>, Denormalized<S1>>,
-  CondNull<P2, DenormalizedNullable<S2>, Denormalized<S2>>,
-  CondNull<P3, DenormalizedNullable<S3>, Denormalized<S3>>,
-  CondNull<P4, DenormalizedNullable<S4>, Denormalized<S4>>,
-  CondNull<P5, DenormalizedNullable<S5>, Denormalized<S5>>,
+  CondNull<P1, SchemaOf<S1>>,
+  CondNull<P2, SchemaOf<S2>>,
+  CondNull<P3, SchemaOf<S3>>,
+  CondNull<P4, SchemaOf<S4>>,
+  CondNull<P5, SchemaOf<S5>>,
 ];
-export default function useResourceNew<
+export default function useResourceLegacy<
   Params extends Readonly<object>,
-  Body extends Readonly<object | string> | void,
   S extends Schema
 >(...args: ResourceArgs<S, Params> | ResourceArgs<S, Params>[]) {
   // this conditional use of hooks is ok as long as the structure of the arguments don't change
